crawler/requirements.txt
# HTTP and parsing
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Image processing
Pillow>=10.0.0

# YAML handling
PyYAML>=6.0.0

# CLI interface
click>=8.1.0

# Date/time handling
python-dateutil>=2.8.0

# Development and testing
pytest>=8.0.0
pytest-cov>=4.0.0
responses>=0.25.0

# Linting
ruff>=0.3.0
